Change behavior for path selection(s)
When one or more directories are chosen, their corresponding file
value(s) will be set to an empty string, and their corresponding
path value(s) will be set to the concatenated full path.

For cases where at least one directory is chosen and the user does
not specify the path output, a warning is thrown to notify the user
of this behavior.